- check http status

+++ src/tools/image.rs
@@ -17,47 +17,54 @@ pub fn optimize_image(img: &mut Image, http: &Http) {
 
     if !local_path.exists() {
         let tmp_path = temp_dir().to_str().unwrap().to_string() + "/" + &img.hash;
+        let tmp_path = Path::new(tmp_path.as_str());
         {
-            let raw: Vec<u8> = http.get_raw(&img.remote_url);
-            let mut tmp = File::create(&tmp_path).expect("could not create file");
-            tmp.write_all(&*raw)
-                .expect("could not write temporary image file");
+            let raw = http.get_raw(&img.remote_url);
+            if raw.is_ok() {
+                let mut tmp = File::create(&tmp_path).expect("could not create file");
+                tmp.write_all(&*raw.unwrap())
+                    .expect("could not write temporary image file");
 
-            tmp.sync_data().expect("file sync failed");
+                tmp.sync_data().expect("file sync failed");
+            }
         }
 
-        let mut c1 = Command::new("convert")
-            .args([
-                "-resize",
-                "300",
-                "-strip",
-                &tmp_path,
-                local_path.to_str().unwrap(),
-            ])
-            .spawn()
-            .expect("could not start jpeg conversion");
-        let mut c2 = Command::new("convert")
-            .args([
-                "-resize",
-                "300",
-                "-strip",
-                "-define",
-                "heic:speed=2",
-                &tmp_path,
-                (LOCAL_DIR.to_owned() + &img.public_avif_url).as_str(),
-            ])
-            .spawn()
-            .expect("could not start avif conversion");
+        if tmp_path.exists() {
+            let mut c1 = Command::new("convert")
+                .args([
+                    "-resize",
+                    "300",
+                    "-strip",
+                    tmp_path.to_str().unwrap(),
+                    local_path.to_str().unwrap(),
+                ])
+                .spawn()
+                .expect("could not start jpeg conversion");
+            let mut c2 = Command::new("convert")
+                .args([
+                    "-resize",
+                    "300",
+                    "-strip",
+                    "-define",
+                    "heic:speed=2",
+                    tmp_path.to_str().unwrap(),
+                    (LOCAL_DIR.to_owned() + &img.public_avif_url).as_str(),
+                ])
+                .spawn()
+                .expect("could not start avif conversion");
 
-        c1.wait().unwrap();
-        c2.wait().unwrap();
+            c1.wait().unwrap();
+            c2.wait().unwrap();
 
-        remove_file(tmp_path).unwrap();
+            remove_file(tmp_path).unwrap();
+        }
     }
 
-    let (width, height) = get_dimension(local_path.to_str().unwrap());
-    if width > 0 && height > 0 {
-        img.set_size(width, height);
+    if local_path.exists() {
+        let (width, height) = get_dimension(local_path.to_str().unwrap());
+        if width > 0 && height > 0 {
+            img.set_size(width, height);
+        }
     }
 }
